code/requirements.txt
Flask=="*"
Flask-RESTful=="*"
Flask-SQLAlchemy=="*"
Flask-JWT-Extended=="*"
Flask-marshmallow=="*"
Flask-Uploads=="*"
authlib=="*"
python-dotenv=="*"
requests=="*"
